feat(initiatives): unify list + detail into master-detail
Stacked on #142.

Folds the standalone /initiatives/[id] page into a master-detail layout
on /initiatives so picking an initiative from the tree shows its full
detail UI in the right pane — no navigation, no scroll-position loss,
no "huge dead band on either side of the form."

Mechanics:
- Extract InitiativeDetailView from /initiatives/[id]/page.tsx into
  src/components/InitiativeDetailView.tsx as a reusable client
  component with `variant: "full" | "pane"`. "full" keeps the
  page-fill wrapper + back-link breadcrumb (used by the standalone
  route); "pane" omits both so the host (PageWithRails main column)
  controls sizing.
- /initiatives/[id]/page.tsx becomes a thin route shell rendering
  InitiativeDetailView in "full" — preserves deep links and the
  "open in focus mode" workflow.
- /initiatives uses PageWithRails: planning tree as left rail
  (w-[26rem]), InitiativeDetailView in main column. Selection is
  URL-backed via ?selected=<id> so back/forward + bookmarks work;
  local state is the source of truth for re-render reliability
  (router.replace was racing the page's own useSearchParams snapshot).
- InitiativeRow trimmed: clicking the title now selects the row
  (highlighted with mc-accent border + bold) instead of toggling an
  inline details accordion. The ExternalLink-to-detail icon is gone —
  the right pane is the detail. Tree is denser (smaller padding /
  badges / counts) so more of the hierarchy fits in the rail.
- PageWithRails.leftRailWidth made configurable so list-style left
  rails can use w-[26rem] vs the default w-64 settings nav.

Verified in preview: tree on the left at 1600px viewport, detail
fills the right; clicking a row updates URL + right pane; standalone
/initiatives/[id] still renders the same UI in full-width mode with
the back-link.
